Clarify benchmark results table
The table showing the results for the different facilities had a column
named "Type", which in the case of the jsonslicer tests corresponded to
the text type of the produced output. To clarify this I've changed the
name to "Output type", and also removed the information redundancy in
the jsonslicer test case names, where the output type was mentioned too.

I have also corrected the output type for the ijson tests, as ijson
always outputs str.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>