Fix comparison with ijson's items functions
The latter always take paths as str, not as bytes, and no internal check
is done (yet) to double-check this. As a result the comparisons with
ijson were bogus, as ijson.items was yielding no results (with python
3.8.2; in python 2.7 they would have yielded the correct results, but
jsonslicer doesn't support 2.7).

To avoid bogus comparisons in the future a new assertion in the common
assertion code now verifies that the expected number of objects were
yielded by the different methods.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>